pyproject.toml
[tool.poetry]
name = "network_live"
version = "0.1.0"
description = ""
authors = ["Ramil Albakov <ralbakov10@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
enm-client-scripting = {path = "../enm_client_scripting-1.20.1-py2.py3-none-any.whl"}
python-dotenv = "^0.19.2"
cx-Oracle = "^8.3.0"
paramiko = "^2.9.2"
defusedxml = "^0.7.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
wemake-python-styleguide = "^0.16.0"

[tool.poetry.scripts]
update-network-live = "network_live.scripts.update_network_live:main"
update-enm-lte = "network_live.scripts.update_enm_lte:main"
update-enm-wcdma = "network_live.scripts.update_enm_wcdma:main"
update-enm-gsm = "network_live.scripts.update_enm_gsm:main"
update-tele2-lte = "network_live.scripts.update_tele2_lte:main"
update-tele2-wcdma = "network_live.scripts.update_tele2_wcdma:main"
update-beeline-huawei = "network_live.scripts.update_beeline_huawei:main"
update-nokia-lte = "network_live.scripts.update_beeline_nokia_lte:main"
update-nokia-wcdma = "network_live.scripts.update_beeline_nokia_wcdma:main"
update-oss-wcdma = "network_live.scripts.update_oss_wcdma:main"
update-huawei-wcdma = "network_live.scripts.update_beeline_huawei_wcdma:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
